requirements.txt
Django==4.2.7
wagtail==5.2.2
django-meta==2.5.0
djangorestframework==3.14.0
Pillow==10.0.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-environ==0.11.2
